Moving Lobby component on his own folder
